backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
easyocr
ffmpeg-python
python-dotenv
sqlalchemy
psycopg2-binary
